backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
scikit-learn>=1.2.0
pandas>=2.0.0
pillow>=9.0.0
numpy>=1.24.0
tqdm
matplotlib
requests
google-genai
python-dotenv
scipy>=1.10.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
chromadb>=0.4.22
pypdf>=4.0.0
sentence-transformers>=2.3.0
